Adopt bull and refactor the relayer codebase
