+++ .nvmrc
@@ -1,1 +1,1 @@
-18.12.1
+20.10.0
