+++ package.json
@@ -23,6 +23,7 @@
   "homepage": "https://github.com/XinFinOrg/XDC-Relayer#readme",
   "repository": "https://github.com/XinFinOrg/XDC-Relayer",
   "devDependencies": {
+    "@types/bull": "^4.10.0",
     "@types/bunyan": "^1.8.8",
     "@types/cron": "^1.7.2",
     "@types/jest": "^29.4.0",
@@ -31,6 +32,7 @@
     "@types/koa__router": "^8.0.4",
     "@types/koa-bodyparser": "^4.3.0",
     "@types/koa-helmet": "^6.0.2",
+    "@types/lodash": "^4.14.202",
     "@types/node": "^18.14.1",
     "@types/node-fetch": "^2.6.9",
     "@typescript-eslint/eslint-plugin": "^6.7.0",
@@ -50,12 +52,14 @@
     "@koa/router": "^10.0.0",
     "agentkeepalive": "^4.3.0",
     "axios": "^1.3.4",
+    "bull": "^4.11.5",
     "bunyan": "^1.8.15",
     "cron": "^1.8.2",
     "dotenv": "^8.2.0",
     "koa": "^2.13.1",
     "koa-bodyparser": "^4.3.0",
     "koa-helmet": "^6.1.0",
+    "lodash": "^4.17.21",
     "node-cache": "^5.1.2",
     "node-fetch": "2",
     "patch-package": "^6.5.1",
